requirements.txt
# Multi-Agent Consensus Framework
# Python 3.11+ required

# Core Dependencies
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1

# LLM Providers
openai>=1.3.0
anthropic>=0.7.0
# xai-sdk>=0.1.0  # Uncomment when available

# HTTP & API
requests>=2.31.0
httpx>=0.25.0
websockets>=12.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
hypothesis>=6.92.0

# Code Quality
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Monitoring & Observability
prometheus-client>=0.19.0
structlog>=23.2.0

# Optional: TTS Integration
# elevenlabs>=0.2.0

# Optional: Nostr Integration
# pynostr>=0.1.0

# Optional: Lightning Network
# bolt11>=0.1.0

# Utilities
tqdm>=4.66.0
rich>=13.7.0
click>=8.1.0
